requirements.txt
flask
Flask-RESTful
requests
Flask-Cors
PyJWT
pymongo
black
redis